fix: parsing parameters

+++ include/septentrio_gnss_driver/communication/settings.hpp
@@ -333,21 +333,21 @@ struct Settings
     //! Whether or not to publish the tf of the localization
     bool publish_tf_ecef;
     //! Minimum period [ms] of the ok rate bound status for pvt topics
-    double monitor_pvt_period_ok_min_ms;
+    uint32_t monitor_pvt_period_ok_min_ms;
     //! Maximum period [ms] of the ok rate bound status for pvt topics
-    double monitor_pvt_period_ok_max_ms;
+    uint32_t monitor_pvt_period_ok_max_ms;
     //! Minimum period [ms] of the warn rate bound status for pvt topics
-    double monitor_pvt_period_warn_min_ms;
+    uint32_t monitor_pvt_period_warn_min_ms;
     //! Maximum period [ms] of the warn rate bound status for pvt topics
-    double monitor_pvt_period_warn_max_ms;
+    uint32_t monitor_pvt_period_warn_max_ms;
     //! Minimum period [ms] of the ok rate bound status for rest topics
-    double monitor_rest_period_ok_min_ms;
+    uint32_t monitor_rest_period_ok_min_ms;
     //! Maximum period [ms] of the ok rate bound status for rest topics
-    double monitor_rest_period_ok_max_ms;
+    uint32_t monitor_rest_period_ok_max_ms;
     //! Minimum period [ms] of the warn rate bound status for rest topics
-    double monitor_rest_period_warn_min_ms;
+    uint32_t monitor_rest_period_warn_min_ms;
     //! Maximum period [ms] of the warn rate bound status for rest topics
-    double monitor_rest_period_warn_max_ms;
+    uint32_t monitor_rest_period_warn_max_ms;
     //! Whether or not to monitor topic rate of the GGA message
     bool monitor_gpgga;
     //! Whether or not to monitor topic rate of the RMC message
